## 변경사항 - 게시판 헤더 추가 - 게시판 로그아웃 추가 - 게시판 리스트 새글, 비밀글 아이콘 추가 - 게시판 수정, 삭제 로그인 확인 기능 추가 ## 추후계획 - 코드 정리 - ui 개선

+++ src/main/webapp/BoardArticleDeleteForm.jsp
@@ -8,7 +8,13 @@
 	}
 	String num = request.getParameter("num");
 	BoardArticleDAO boardArticle = new BoardArticleDAO();
-	boardArticle.deleteArticle(num);
+	BoardArticleVO board = boardArticle.readArticle(num);	
+	if(Objects.equals(board.getWriter(),session.getAttribute("user_id")) || Objects.equals(session.getAttribute("user_id"), "admin")){
+		boardArticle.deleteArticle(num);	
+	} else {
+		out.println("<script>alert('작성자 혹은 관리자가 아닙니다.'); history.back();</script>");
+	}
+	
 
 %>
 <script>
